+++ LibCarla/source/carla/trafficmanager/SimpleWaypoint.cpp
@@ -45,11 +45,11 @@ namespace traffic_manager {
     return waypoint->GetTransform().rotation.GetForwardVector();
   }
 
-  uint SimpleWaypoint::SetNextWaypoint(const std::vector<SimpleWaypointPtr> &waypoints) {
+  uint64_t SimpleWaypoint::SetNextWaypoint(const std::vector<SimpleWaypointPtr> &waypoints) {
     for (auto &simple_waypoint: waypoints) {
       next_waypoints.push_back(simple_waypoint);
     }
-    return static_cast<uint>(waypoints.size());
+    return static_cast<uint64_t>(waypoints.size());
   }
 
   void SimpleWaypoint::SetLeftWaypoint(SimpleWaypointPtr _waypoint) {
